t1863: extract email_shared.py — deduplicate shared code between email-summary.py and entity-extraction.py
- Create .agents/scripts/email_shared.py with shared utilities:
  extract_body, extract_frontmatter, yaml_escape, check_ollama, get_ollama_model
- Remove duplicate implementations from both files (15+ duplicated lines)
- Both files now import from email_shared, eliminating the qlty duplication smell
- No function exceeds complexity 12 (highest: _parse_summary_response=9, extract_entities_spacy=10)
- Both scripts pass --help and functional tests unchanged

Closes #16059
